Move scripts to their own library
