+++ src/core/console/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(${PROJECT_CORE_LIB} PRIVATE
     Console.cc
     ConsoleBindingManager.cc
     ConsoleCoreCommands.cc
+    ConsoleTracy.cc
 )
 
 target_precompile_headers(${PROJECT_CORE_LIB} PUBLIC
