+++ src/scripts/prefabs/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(${PROJECT_SCRIPTS_LIB} PRIVATE
+    GltfPrefab.cc
+)
