+++ src/scripts/scripts/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${PROJECT_SCRIPTS_LIB} PRIVATE
+    ArticulatingArm.cc
+    LightScripts.cc
+    MiscScripts.cc
+)
